moved inputValue hook to Name component
