+++ src/components/settings/NameInput.js
@@ -1,4 +1,11 @@
+import { useState } from "react";
+
 const NameInput = (props) => {
+  const [inputValue, setInputValue] = useState("");
+
+  function handleInputValue(event) {
+    setInputValue(event.target.value);
+  }
   return (
     <div className="settings-container flexbox">
       <div>
@@ -7,11 +14,17 @@ const NameInput = (props) => {
           id="name-input"
           type="text"
           name="name"
-          onChange={props.onChange}
-          value={props.value}
+          onChange={handleInputValue}
+          value={inputValue}
         />
       </div>
-      <button className="small-button" onClick={props.saveName}>
+      <button
+        className="small-button"
+        onClick={() => {
+          props.saveName(inputValue);
+          setInputValue("");
+        }}
+      >
         Spara
       </button>
     </div>
